workflow-orchestrator/src/requirements.txt
boto3>=1.34.0
langgraph==0.4.10
langchain-core>=0.2.0
# Requests library
requests>=2.31.0
# For schema validation
jsonschema>=4.21.0
# For YAML parsing
PyYAML>=6.0
# For JSONPath parsing
jsonpath-ng>=1.6.0
# For flexible date parsing
python-dateutil>=2.8.0